Remove band-aid
-Upgrade MCSwipeTableViewCell

Swiping to view comments then clicking to view article immediately
(webpage) before comment can finish segue causes the app to crash. This
happens very rarely.

[self.tableView setAllowsSelection:NO];
was included in the old version when swiping began and
[self.tableView performSelector:@selector(setAllowsSelection:)
withObject:[NSNumber numberWithBool:YES] afterDelay:0.5];

allows tableView selection only when swiping ended to prevent freezing.
This fix is now removed because it is causing complete freezing when
not letting the swipe cell slap shut on 64 bit iOS.

swipeTableViewCellDidEndSwiping is not called for some reason.